Create odel Inferencing on the Web to validate offline accuracy
10.	Model Inferencing on the Web to validate offline accuracy
The BiLSTM-Informer model was deployed on the web using Orender following a structured workflow designed to ensure smooth integration, efficient inference, and seamless user interaction. First, the trained model was saved in an appropriate format (.pb) to maintain compatibility with TensorFlow's deployment ecosystem. The model and essential scripts for inference, preprocessing, and API integration were stored in a GitHub repository, ensuring version control and accessibility. A requirements.txt file was created to define all dependencies, enabling smooth replication of the development environment.